+++ packages/api/src/settings/transformers.ts
@@ -1,8 +1,16 @@
 import { SettingDocument, SettingDto } from './types';
 
 const settingDocumentToDto = (setting: SettingDocument): SettingDto => {
-  const { _id, key, value, type, label, description } = setting;
-  return { _id, key, value, type, label, description };
+  const { _id, key, value, valueRealized, type, label, description } = setting;
+  return {
+    _id,
+    key,
+    value,
+    valueRealized,
+    type,
+    label,
+    description,
+  };
 };
 
 export const settingListTransformer = (