+++ packages/api/src/settings/validators.ts
@@ -1,12 +1,15 @@
 import { SettingDocument } from './types';
+import { PeriodSettingDocument } from '@periodsettings/types';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 function isNumeric(num: any): Boolean {
   return !isNaN(num);
 }
 
-export function fieldTypeValidator(this: SettingDocument): Boolean {
-  if (this.type === 'Number') {
+export function fieldTypeValidator(
+  this: SettingDocument | PeriodSettingDocument
+): Boolean {
+  if (this.type === 'Float' || this.type === 'Integer') {
     return isNumeric(this.value);
   }
 
@@ -22,7 +25,7 @@ export function fieldTypeValidator(this: SettingDocument): Boolean {
     return this.value === 'true' || this.value === 'false';
   }
 
-  if (this.type === 'List') {
+  if (this.type === 'IntegerList') {
     let valid = true;
     let previous = 0;
     const valueArray = this.value.split(',').map((item) => item.trim());