+++ packages/frontend/src/pages/Periods/Details/components/Details.tsx
@@ -6,7 +6,9 @@ import {
   useAssignQuantifiers,
   useClosePeriod,
   useExportPraise,
+  useVerifyQuantifierPoolSize,
 } from '@/model/periods';
+import { AllPeriodSettings } from '@/model/periodsettings';
 import { AllQuantifierUsers } from '@/model/users';
 import { formatDate } from '@/utils/date';
 import { saveLocalFile } from '@/utils/file';
@@ -35,9 +37,14 @@ const PeriodDetails = (): JSX.Element | null => {
   const allQuantifiers = useRecoilValue(AllQuantifierUsers);
   const { periodId } = useParams<PeriodPageParams>();
   const period = useRecoilValue(SinglePeriod(periodId));
+  const periodsettings = useRecoilValue(AllPeriodSettings(periodId));
   const isAdmin = useRecoilValue(HasRole(ROLE_ADMIN));
   const { exportPraise } = useExportPraise();
   const history = useHistory();
+  const poolRequirements = useVerifyQuantifierPoolSize(
+    periodId,
+    JSON.stringify(periodsettings)
+  );
 
   const assignDialogRef = React.useRef(null);
   const closeDialogRef = React.useRef(null);
@@ -185,6 +192,7 @@ const PeriodDetails = (): JSX.Element | null => {
             <PeriodAssignDialog
               onClose={(): void => setIsAssignDialogOpen(false)}
               onAssign={(): void => handleAssign()}
+              poolRequirements={poolRequirements}
             />
           </div>
         </Dialog>