+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Modify settings architecture so Periods have their own unique settings. New periods are created with the "global" settings as default values. #138 #116
+
 ## [0.2.0] - 2022-03-31
 
 ### Added